Do not create the /usr/lib64/xorg/modules/nvidia-%{version} symlink and remove redundant Optimus configuration.
On systems prior to Fedora 25 the modules directory is specified in the file /etc/X11/xorg.conf.d/99-nvidia.conf, which is always loaded even if a user has a custom xorg.conf. On Fedora 25+, if the driver is installed and is not actually used (fallback scenario) this makes the X server to always load the Nvidia extension.

On Fedora 25+ the OutputClass configuration for the i915 driver is alreay at this default for Intel gen4+ systems, on Intel gen3 systems it prevents the usage of Glamor and there are no gen3 Optimus laptops.